integration: extract metadata/contents.rs
These are tests that simply check that driver keeps metadata of some
kind. As these tests are similar in their concept, I believe it makes
sense to have them all in one file.
