+++ scylla/tests/integration/metadata/mod.rs
@@ -1,1 +1,2 @@
 mod configuration;
+mod contents;
